+++ multimodal/src/autogluon/multimodal/optimization/__init__.py
@@ -1,4 +1,4 @@
-from . import lit_module, utils
+# from . import lit_module, utils
 from .lit_module import LitModule
 from .lit_distiller import DistillerLitModule
 from .lit_ner import NerLitModule
